add page.rs
